# feat: Batch GitHub API calls for 20x faster cleanup