--- a/admin/docs/enhancements/batch-github-api-calls.md
+++ b/admin/docs/enhancements/batch-github-api-calls.md
@@ -313,13 +313,32 @@ time ./scripts/workflow-helper.sh clean
 
 ## 📝 Implementation Log
 
-### **October 3, 2025**
-- ✅ Created enhancement documentation
-- 🔄 Started Phase 1: jq dependency check
-- 📋 Next: Implement `gf_check_jq()` function
+### **October 3, 2025 - Session 1**
+- ✅ Created enhancement documentation (351 lines)
+- ✅ **Phase 1 COMPLETE**: jq dependency check
+  - Implemented `gf_check_jq()` function in `git-flow-utils.sh`
+  - Added to exports for use in other scripts
+  - Tested and verified working
+  
+- ✅ **Phase 2 COMPLETE**: Batched API call functions
+  - Implemented `get_merged_branches_batched()` function
+  - Implemented `get_merged_branches_individual()` fallback
+  - Comprehensive error handling
+  - Tested and verified working
+
+- ✅ **Phase 3 COMPLETE**: Integration with cleanup command
+  - Updated local branch cleanup logic
+  - Updated remote branch cleanup logic
+  - Automatic jq detection and fallback
+  - User-friendly status messages
+
+- 📋 **Phase 4 PENDING**: Documentation updates
+  - Update README.md with jq dependency
+  - Update DEVELOPMENT.md with performance notes
+  - Update enhancement tracking
 
 ---
 
-**Status**: Phase 1 in progress  
-**Next Steps**: Complete jq check, then implement batched API call  
-**Expected Completion**: October 3-4, 2025
+**Status**: Phases 1-3 complete (75% done)  
+**Next Steps**: Phase 4 - Documentation updates  
+**Expected Completion**: October 3, 2025 (today!)